Iniciando Student Controller

+++ src/main/java/com/crednovo/desafio/repository/TeacherRepository.java
@@ -3,7 +3,7 @@ package com.crednovo.desafio.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.crednovo.desafio.models.Teacher;
 
-public interface TeacherRepository extends JpaRepository <Teacher, Integer> {
+public interface TeacherRepository extends JpaRepository <Teacher, Long> {
 
 }
